+++ NEWS.md
@@ -2,6 +2,7 @@ annolite 0.0.1.9001
 ===================
 
 * The `annotate()` function will check whether the parent directory of a file stated via argument `file` exists and throws a warning if not (#6).
+* The package *testthat*  is now a suggested package and there is a minimal testsuite.
 
 
 annolite 0.0.1