+++ NEWS.md
@@ -1,5 +1,5 @@
-annolite 0.0.1.9001
-===================
+annolite 0.0.1.9001 - 0.0.1.9002
+================================
 
 * The `annotate()` function will check whether the parent directory of a file stated via argument `file` exists and throws a warning if not (#6).
 * The package *testthat*  is now a suggested package and there is a minimal testsuite.
@@ -9,6 +9,8 @@ annolite 0.0.1.9001
 * The `annolite()` function coerce the input argument `annotations` to a `annotationstable` object, running a validity check if it is a `data.frame` (#7).
 * The *annolite* did not work as intended if argument `crosstalk`  was `FALSE` due to a misplaced JavaScript check whether the crosstalk library is present. The text is more robust now and the widget works as intended (#10).
 * The JavaScript part of the package will handle situations correctly if the end position of a text selection is before the start position / annotation from right to left (#11).
+* Deleting annotations did not work correctly, the annotation was removed even if user cancelled the action - fixed (#14).
+* The example for the `annotate()` function still used a speech of Volker Kauder. Now it is the speech of Kofi Annan on the Millennium Declaration.
 
 
 annolite 0.0.1